requirements.txt
asgiref==3.5.2
Brotli==1.0.9
cffi==1.15.1
cssselect2==0.6.0
Django==4.0.6
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-serializer-extensions==2.0.1
fonttools==4.34.4
hashids==1.3.1
html5lib==1.1
Pillow==9.2.0
pycparser==2.21
pydyf==0.2.0
pyphen==0.12.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.1
tzdata==2022.1
weasyprint==56.0
webencodings==0.5.1
zopfli==0.2.1
